pyproject.toml
[project]
name = "kaggle"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "ruff>=0.7.4",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "transformers>=4.46.3",
    "lightgbm>=4.5.0",
    "catboost>=1.2.7",
    "pytorch-lightning>=2.4.0",
    "polars>=1.13.1",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
    "scikit-learn>=1.5.2",
    "notebook<=7.2.2",
    "ipykernel>=6.29.5",
    "opencv-python>=4.10.0.84",
    "matplotlib-venn>=1.1.1",
    "pyarrow>=18.0.0",
    "hydra-core>=1.3.2",
    "hydra-colorlog>=1.2.0",
    "timm>=1.0.11",
    "albumentations>=1.4.21",
    "wandb>=0.18.7",
    "xgboost>=2.1.2",
    "simdkalman>=1.0.4",
    "ultralytics>=8.3.36",
    "lapx>=0.5.11",
    "huggingface-hub>=0.26.2",
    "gradio-imageslider>=0.0.20",
    "gradio>=5.6.0",
    "xformers>=0.0.28.post3",
    "bitsandbytes>=0.44.1",
    "einops>=0.8.0",
    "accelerate>=1.1.1",
    "sentencepiece>=0.2.0",
    "flash-attention>=1.0.0",
    "schedulefree>=1.4",
]
readme = "README.md"
requires-python = "==3.11.9"


[tool.uv.sources]
layerone-ml-utils = { path = "../libs/layerone-ml-utils", editable = true }
torch = [
    { index = "pytorch-cu121", marker = "platform_system == 'Linux'" },
]
torchvision = [
    { index = "pytorch-cu121", marker = "platform_system == 'Linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.mypy]
check_untyped_defs = true
disallow_any_decorated = false
disallow_any_generics = true
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = false
disallow_untyped_calls = false
disallow_untyped_decorators = false
disallow_untyped_defs = false
ignore_errors = false
ignore_missing_imports = true
namespace_packages = true
no_implicit_optional = true
python_version = "3.10"
show_error_codes = true
strict_equality = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = false

exclude = [".venv", "output"]

[tool.pytest.ini_options]
filterwarnings = []

[tool.coverage.run]
omit = ["tests/*"]
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
ignore = [
    "E402", # Module level import not at top of file
    "E501", # Line too long
    "E203", # whitespace before ':' (blackと競合する)
]
select = ["F", "E", "W", "I", "B"]
exclude = [".venv", "output"]

[tool.ruff.lint.isort]
known-first-party = ["components", "modules"]

